backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-pptx
requests
pydantic
google-cloud-aiplatform
google-generativeai